Refactor the code to match master branch
