+++ build.sbt
@@ -72,8 +72,6 @@ lazy val IntegrationTest = config("it") extend Test
 lazy val vlm = project
   .dependsOn(testkit % Test)
   .settings(commonSettings)
-  .configs(IntegrationTest)
-  .settings(Defaults.itSettings)
   .settings(
     organization := "com.azavea.geotrellis",
     name := "geotrellis-contrib-vlm",
@@ -88,25 +86,20 @@ lazy val vlm = project
       sparkSQL % Test,
       geotrellisSparkTestKit % Test,
       scalatest % Test,
-      gdal % Test
+      gdalBindings % Test
     ),
     Test / fork := true,
     Test / parallelExecution := false,
     Test / testOptions += Tests.Argument("-oDF")
   )
-  .settings(
-    initialCommands in console :=
-      """
-        |import geotrellis.contrib.vlm._
-        |import geotrellis.contrib.vlm.geotiff._
-        |import geotrellis.contrib.vlm.gdal._
-      """.stripMargin
-  )
+  
 
 lazy val gdal = project
   .dependsOn(testkit % Test)
   .dependsOn(vlm)
   .settings(commonSettings)
+  .configs(IntegrationTest)
+  .settings(Defaults.itSettings)
   .settings(
     organization := "com.azavea.geotrellis",
     name := "geotrellis-contrib-gdal",
