+++ gdal/src/main/scala/geotrellis/contrib/vlm/gdal/GDALRasterSource.scala
@@ -19,7 +19,6 @@ package geotrellis.contrib.vlm.gdal
 import geotrellis.contrib.vlm.TargetCellType
 import geotrellis.raster.resample.ResampleMethod
 
-
 case class GDALRasterSource(
   uri: String,
   options: GDALWarpOptions = GDALWarpOptions(),
