+++ gdal/src/main/scala/geotrellis/contrib/vlm/gdal/GDALResampleRasterSource.scala
@@ -25,7 +25,6 @@ import com.azavea.gdal.GDALWarp
 
 import cats.syntax.option._
 
-
 case class GDALResampleRasterSource(
   uri: String,
   resampleGrid: ResampleGrid,
