+++ gdal/src/main/scala/geotrellis/contrib/vlm/gdal/package.scala
@@ -23,7 +23,6 @@ import geotrellis.vector.Extent
 
 import cats.syntax.option._
 
-
 package object gdal {
 
   implicit class GDALRasterExtentMethods(val self: RasterExtent) {
