+++ gdal/src/test/resources/application.conf
@@ -1,0 +1,18 @@
+gdal {
+  settings {
+    options {
+      GDAL_MAX_DATASET_POOL_SIZE = "100"
+      VRT_SHARED_SOURCE          = "0"
+      CPL_DEBUG                  = "OFF"
+      GDAL_HTTP_MAX_RETRY        = "10"
+    }
+    useExceptions: true
+  }
+  cache {
+    # maximumSize: 1000
+    enableDefaultRemovalListener: true
+    valuesType: Weak
+    enabled: true
+    withShutdownHook: true
+  }
+}
