+++ gdal/src/test/scala/geotrellis/contrib/vlm/gdal/GDALConvertedRasterSourceSpec.scala
@@ -19,19 +19,9 @@ package geotrellis.contrib.vlm.gdal
 import geotrellis.contrib.vlm._
 import geotrellis.raster._
 import geotrellis.raster.io.geotiff.reader.GeoTiffReader
-import geotrellis.raster.io.geotiff.AutoHigherResolution
-import geotrellis.raster.resample._
 import geotrellis.raster.testkit._
-import geotrellis.vector._
-import geotrellis.spark._
-import geotrellis.util._
-
-import com.azavea.gdal.GDALWarp
 
 import org.scalatest._
-
-import java.net.MalformedURLException
-
 
 class GDALConvertedRasterSourceSpec extends FunSpec with RasterMatchers with BetterRasterMatchers with GivenWhenThen {
 
@@ -41,21 +31,21 @@ class GDALConvertedRasterSourceSpec extends FunSpec with RasterMatchers with Bet
   val url2 = Resource.path("img/byte-tile.tiff")
   val uri2 = s"file://$url2"
 
-  val source: GDALRasterSource = GDALRasterSource(url)
-  val byteSource: GDALRasterSource = GDALRasterSource(url2)
+  lazy val source: GDALRasterSource = GDALRasterSource(url)
+  lazy val byteSource: GDALRasterSource = GDALRasterSource(url2)
 
-  val expectedRaster: Raster[MultibandTile] =
+  lazy val expectedRaster: Raster[MultibandTile] =
     GeoTiffReader
       .readMultiband(url, streaming = false)
       .raster
 
-  val expectedRaster2: Raster[MultibandTile] =
+  lazy val expectedRaster2: Raster[MultibandTile] =
     GeoTiffReader
       .readMultiband(url2, streaming = false)
       .raster
 
-  val targetExtent = expectedRaster.extent
-  val targetExtent2 = expectedRaster2.extent
+  lazy val targetExtent = expectedRaster.extent
+  lazy val targetExtent2 = expectedRaster2.extent
 
   /** Note:
    *  Many of these tests have a threshold of 1.0. The reason for this large value
