+++ gdal/version.sbt
@@ -1,0 +1,1 @@
+version in ThisBuild := "0.10.1-SNAPSHOT"
