+++ vlm/src/test/scala/geotrellis/contrib/vlm/spark/RasterSourceRDDSpec.scala
@@ -30,33 +30,25 @@ import geotrellis.spark._
 import geotrellis.spark.io.hadoop._
 import geotrellis.spark.testkit._
 import geotrellis.spark.tiling._
-
-import geotrellis.contrib.vlm._
-
 import org.apache.spark.rdd.RDD
 import org.scalatest.Inspectors._
 import org.scalatest._
-
 import spire.syntax.cfor._
 
-import com.azavea.gdal.GDALWarp
-
 import scala.concurrent.ExecutionContext
-
 
 class RasterSourceRDDSpec extends FunSpec with TestEnvironment with BetterRasterMatchers with BeforeAndAfterAll {
   val filePath = s"${new File("").getAbsolutePath()}/src/test/resources/img/aspect-tiled.tif"
   def filePathByIndex(i: Int): String = s"${new File("").getAbsolutePath()}/src/test/resources/img/aspect-tiled-$i.tif"
   val uri = s"file://$filePath"
-  val rasterSource = GeoTiffRasterSource(uri)
-  val targetCRS = CRS.fromEpsgCode(3857)
-  val scheme = ZoomedLayoutScheme(targetCRS)
-  val layout = scheme.levelForZoom(13).layout
+  lazy val rasterSource = GeoTiffRasterSource(uri)
+  lazy val targetCRS = CRS.fromEpsgCode(3857)
+  lazy val scheme = ZoomedLayoutScheme(targetCRS)
+  lazy val layout = scheme.levelForZoom(13).layout
 
-  val reprojectedSource = rasterSource.reprojectToGrid(targetCRS, layout)
+  lazy val reprojectedSource = rasterSource.reprojectToGrid(targetCRS, layout)
 
   describe("reading in GeoTiffs as RDDs") {
-
     it("should have the right number of tiles") {
       val expectedKeys =
         layout
@@ -109,10 +101,10 @@ class RasterSourceRDDSpec extends FunSpec with TestEnvironment with BetterRaster
     }
 
     def assertRDDLayersEqual(
-      expected: MultibandTileLayerRDD[SpatialKey],
-      actual: MultibandTileLayerRDD[SpatialKey],
-      matchRasters: Boolean = false
-    ): Unit = {
+                              expected: MultibandTileLayerRDD[SpatialKey],
+                              actual: MultibandTileLayerRDD[SpatialKey],
+                              matchRasters: Boolean = false
+                            ): Unit = {
       val joinedRDD = expected.filter { case (_, t) => !t.band(0).isNoDataTile }.leftOuterJoin(actual)
 
       joinedRDD.collect().foreach { case (key, (expected, actualTile)) =>
@@ -191,120 +183,9 @@ class RasterSourceRDDSpec extends FunSpec with TestEnvironment with BetterRaster
 
       assertRDDLayersEqual(reprojectedExpectedRDD, tiledSource)
     }
-
-    describe("GDALRasterSource") {
-      val expectedFilePath = s"${new File("").getAbsolutePath()}/src/test/resources/img/aspect-tiled-near-merc-rdd.tif"
-
-      it("should reproduce tileToLayout") {
-        val rasterSource = GDALRasterSource(filePath)
-
-        // This should be the same as result of .tileToLayout(md.layout)
-        val rasterSourceRDD: MultibandTileLayerRDD[SpatialKey] = RasterSourceRDD(rasterSource, md.layout)
-        // Complete the reprojection
-        val reprojectedSource = rasterSourceRDD.reproject(targetCRS, layout)._2
-
-        assertRDDLayersEqual(reprojectedExpectedRDD, reprojectedSource, true)
-      }
-
-      it("should reproduce tileToLayout followed by reproject GDAL") {
-        val expectedRasterSource = GDALRasterSource(expectedFilePath)
-        val reprojectedExpectedRDDGDAL: MultibandTileLayerRDD[SpatialKey] = RasterSourceRDD(expectedRasterSource, layout)
-        val rasterSource = GDALRasterSource(filePath)
-        val reprojectedRasterSource = rasterSource.reprojectToGrid(targetCRS, layout)
-
-        // This should be the same as .tileToLayout(md.layout).reproject(crs, layout)
-        val reprojectedSourceRDD: MultibandTileLayerRDD[SpatialKey] = RasterSourceRDD(reprojectedRasterSource, layout)
-
-        assertRDDLayersEqual(reprojectedExpectedRDDGDAL, reprojectedSourceRDD, true)
-      }
-
-      def parellSpec(n: Int = 1000)(implicit cs: ContextShift[IO]): List[RasterSource] = {
-        println(java.lang.Thread.activeCount())
-
-        /** Functions to trigger Datasets computation */
-        def ltsWithDatasetsTriggered(lts: LayoutTileSource): LayoutTileSource = { rsWithDatasetsTriggered(lts.source); lts }
-        def rsWithDatasetsTriggered(rs: RasterSource): RasterSource = {
-          val brs = rs.asInstanceOf[GDALBaseRasterSource]
-          brs.dataset.rasterExtent
-          brs.dataset.rasterExtent(GDALWarp.SOURCE)
-          rs
-        }
-
-        /** Do smth usual with the original RasterSource to force VRTs allocation */
-        def reprojRS(i: Int): LayoutTileSource =
-          ltsWithDatasetsTriggered(
-            rsWithDatasetsTriggered(
-              rsWithDatasetsTriggered(GDALRasterSource(filePathByIndex(i)))
-                .reprojectToGrid(targetCRS, layout)
-            ).tileToLayout(layout)
-          )
-
-        /** Simulate possible RF backsplash calls */
-        def dirtyCalls(rs: RasterSource): RasterSource = {
-          val ds = rs.asInstanceOf[GDALBaseRasterSource].dataset
-          ds.rasterExtent
-          ds.crs
-          ds.cellSize
-          ds.extent
-          rs
-        }
-
-        val res = (1 to n).toList.flatMap { _ =>
-          (0 to 4).flatMap { i =>
-            List(IO {
-              // println(Thread.currentThread().getName())
-              // Thread.sleep((Math.random() * 100).toLong)
-              val lts = reprojRS(i)
-              lts.readAll(lts.keys.take(10).toIterator)
-              reprojRS(i).source.resolutions
-
-              dirtyCalls(reprojRS(i).source)
-            }, IO {
-              // println(Thread.currentThread().getName())
-              // Thread.sleep((Math.random() * 100).toLong)
-              val lts = reprojRS(i)
-              lts.readAll(lts.keys.take(10).toIterator)
-              reprojRS(i).source.resolutions
-
-              dirtyCalls(reprojRS(i).source)
-            }, IO {
-              // println(Thread.currentThread().getName())
-              // Thread.sleep((Math.random() * 100).toLong)
-              val lts = reprojRS(i)
-              lts.readAll(lts.keys.take(10).toIterator)
-              reprojRS(i).source.resolutions
-
-              dirtyCalls(reprojRS(i).source)
-            })
-          }
-        }.parSequence.unsafeRunSync
-
-        println(java.lang.Thread.activeCount())
-
-        res
-      }
-
-      it(s"should not fail on parallelization with a fork join pool") {
-        val i = 1000
-        implicit val cs = IO.contextShift(ExecutionContext.global)
-
-        val res = parellSpec(i)
-      }
-
-      it(s"should not fail on parallelization with a fixed thread pool") {
-        val i = 1000
-        val n = 200
-        val pool = Executors.newFixedThreadPool(n)
-        val ec = ExecutionContext.fromExecutor(pool)
-        implicit val cs = IO.contextShift(ec)
-
-        val res = parellSpec(i)
-      }
-    }
   }
 
   describe("RasterSourceRDD.read") {
-
     val floatingScheme = FloatingLayoutScheme(500, 270)
     val floatingLayout = floatingScheme.levelFor(rasterSource.extent, rasterSource.cellSize).layout
