+++ vlm/src/test/scala/geotrellis/contrib/vlm/spark/RasterSummarySpec.scala
@@ -24,14 +24,9 @@ import geotrellis.raster.resample.Bilinear
 import geotrellis.spark._
 import geotrellis.spark.testkit._
 import geotrellis.spark.tiling._
-import geotrellis.vector.Extent
-
-import com.azavea.gdal.GDALWarp
 
 import org.apache.spark.rdd._
 import org.scalatest._
-import spire.syntax.cfor._
-
 
 class RasterSummarySpec extends FunSpec with TestEnvironment with BetterRasterMatchers with GivenWhenThen {
